backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
Pillow==9.1.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.3.0
pytz==2022.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.4.0
